Lab7/requirements.txt
lxml==5.3.0
mpmath==1.3.0
pyasn1==0.6.1
python-docx==1.1.2
rsa==4.9
sympy==1.13.3
typing_extensions==4.12.2
